Split FileSystem Commands and Direct into partial class files
Refactor monolithic Direct.cs and Commands.cs into focused partial class files:

Direct/:
- Direct.GetContent.cs: File reading operations
- Direct.GetChildItem.cs: Directory listing operations
- Direct.GetLocation.cs: Get current working directory
- Direct.SetLocation.cs: Change current working directory
- Direct.RemoveItem.cs: File/directory deletion with force support

Commands/:
- Commands.GetContent.cs: Shell-style file reading
- Commands.GetChildItem.cs: Shell-style directory listing
- Commands.GetLocation.cs: Shell-style get working directory
- Commands.SetLocation.cs: Shell-style change working directory
- Commands.RemoveItem.cs: Shell-style file/directory deletion

This establishes a scalable pattern for adding future file operations (CopyItem, MoveItem, NewItem, etc.) from Task 020.

All tests passing.
